requirements.txt
requests
bs4
ndp_ep
pandas
msgpack
blosc
kafka-python
confluent-kafka
python-dotenv
scidx_streaming
matplotlib
lz4